refactor flowDE

+++ src/Utils.jl
@@ -14,11 +14,11 @@ function displayIteration(i, residual, itlinear = 0, lastRow = false)
 	end
 end
 
-@inline _displayLine(i::Int, residual::Real, itlinear::Tuple{Int64, Int64}) = @printf("|%8d     │ %16.4e     │ (%4d, %4d)   |\n", i, residual, itlinear[1], itlinear[2])
+@inline _displayLine(i::Int, residual::Real, itlinear::Tuple{Int, Int}) = @printf("|%8d     │ %16.4e     │ (%4d, %4d)   |\n", i, residual, itlinear[1], itlinear[2])
 
-@inline _displayLine(i::Int, residual::Real, itlinear::Int64) = @printf("│%8d     │ %16.4e     │ %8d       │\n", i, residual, itlinear)
+@inline _displayLine(i::Int, residual::Real, itlinear::Int) = @printf("│%8d     │ %16.4e     │ %8d       │\n", i, residual, itlinear)
 
-@inline _displayLine(i::Int, residual::Nothing, itlinear::Tuple{Int64, Int64}) = @printf("│%8d     │                      │ (%4d, %4d)   │\n", i, itlinear[1], itlinear[2])
+@inline _displayLine(i::Int, residual::Nothing, itlinear::Tuple{Int, Int}) = @printf("│%8d     │                      │ (%4d, %4d)   │\n", i, itlinear[1], itlinear[2])
 ####################################################################################################
 function computeEigenvalues(it::ContIterable, u0, par, nev = it.contParams.nev; kwargs...)
 	return it.contParams.newtonOptions.eigsolver(it.J(u0, par), nev; kwargs...)
